+++ packages/blitz-auth/src/server/auth-sessions.ts
@@ -2,17 +2,13 @@ import { fromBase64, toBase64 } from "b64-lite"
 import cookie from "cookie"
 import { IncomingMessage, ServerResponse } from "http"
 import { sign as jwtSign, verify as jwtVerify } from "jsonwebtoken"
-import { assert, isLocalhost, getCookieParser } from "blitz"
+import { assert, isLocalhost, getCookieParser, isPast, differenceInMinutes, addYears, addMinutes, } from "blitz"
 import { MiddlewareResponse } from "@blitzjs/next"
 import type { NextApiRequest, NextApiResponse } from "@blitzjs/next"
 import {
   AuthenticationError,
   AuthorizationError,
   CSRFTokenMismatchError,
-  addMinutes,
-  addYears,
-  differenceInMinutes,
-  isPast,
   EmptyPublicData,
   PublicData,
   IsAuthorizedArgs,
@@ -195,7 +191,7 @@ export class SessionContextClass implements SessionContext {
   $isAuthorized(...args: IsAuthorizedArgs) {
     if (!this.userId) return false
 
-    return global.sessionConfig.isAuthorized({ ctx: this._res.blitzCtx, args })
+    return (global as any).sessionConfig.isAuthorized({ ctx: this._res.blitzCtx, args })
   }
 
   async $create(publicData: PublicData, privateData?: Record<any, any>) {
@@ -402,9 +398,9 @@ const setSessionCookie = (
     cookie.serialize(COOKIE_SESSION_TOKEN(), sessionToken, {
       path: "/",
       httpOnly: true,
-      secure: global.sessionConfig.secureCookies && !isLocalhost(req),
-      sameSite: global.sessionConfig.sameSite,
-      domain: global.sessionConfig.domain,
+      secure: (global as any).sessionConfig.secureCookies && !isLocalhost(req),
+      sameSite: (global as any).sessionConfig.sameSite,
+      domain: (global as any).sessionConfig.domain,
       expires: expiresAt,
     }),
   )
@@ -421,9 +417,9 @@ const setAnonymousSessionCookie = (
     cookie.serialize(COOKIE_ANONYMOUS_SESSION_TOKEN(), token, {
       path: "/",
       httpOnly: true,
-      secure: global.sessionConfig.secureCookies && !isLocalhost(req),
-      sameSite: global.sessionConfig.sameSite,
-      domain: global.sessionConfig.domain,
+      secure: (global as any).sessionConfig.secureCookies && !isLocalhost(req),
+      sameSite: (global as any).sessionConfig.sameSite,
+      domain: (global as any).sessionConfig.domain,
       expires: expiresAt,
     }),
   )
@@ -441,9 +437,9 @@ const setCSRFCookie = (
     res,
     cookie.serialize(COOKIE_CSRF_TOKEN(), antiCSRFToken, {
       path: "/",
-      secure: global.sessionConfig.secureCookies && !isLocalhost(req),
-      sameSite: global.sessionConfig.sameSite,
-      domain: global.sessionConfig.domain,
+      secure: (global as any).sessionConfig.secureCookies && !isLocalhost(req),
+      sameSite: (global as any).sessionConfig.sameSite,
+      domain: (global as any).sessionConfig.domain,
       expires: expiresAt,
     }),
   )
@@ -460,9 +456,9 @@ const setPublicDataCookie = (
     res,
     cookie.serialize(COOKIE_PUBLIC_DATA_TOKEN(), publicDataToken, {
       path: "/",
-      secure: global.sessionConfig.secureCookies && !isLocalhost(req),
-      sameSite: global.sessionConfig.sameSite,
-      domain: global.sessionConfig.domain,
+      secure: (global as any).sessionConfig.secureCookies && !isLocalhost(req),
+      sameSite: (global as any).sessionConfig.sameSite,
+      domain: (global as any).sessionConfig.domain,
       expires: expiresAt,
     }),
   )
@@ -500,8 +496,8 @@ async function getSessionKernel(
       )
       return null
     }
-    debug("global session config", global.sessionConfig)
-    const persistedSession = await global.sessionConfig.getSession(handle)
+    debug("(global as any) session config", (global as any).sessionConfig)
+    const persistedSession = await (global as any).sessionConfig.getSession(handle)
     if (!persistedSession) {
       debug("Session not found in DB")
       return null
@@ -552,7 +548,7 @@ async function getSessionKernel(
       const hasQuarterExpiryTimePassed =
         persistedSession.expiresAt &&
         differenceInMinutes(persistedSession.expiresAt, new Date()) <
-        0.75 * (global.sessionConfig.sessionExpiryMinutes as number)
+        0.75 * ((global as any).sessionConfig.sessionExpiryMinutes as number)
 
       if (hasQuarterExpiryTimePassed) {
         debug("quarter expiry time has passed")
@@ -674,7 +670,7 @@ async function createNewSession(
       antiCSRFToken,
       anonymousSessionToken,
     }
-  } else if (global.sessionConfig.method === "essential") {
+  } else if ((global as any).sessionConfig.method === "essential") {
     debug("Creating new session")
     const newPublicData: PublicData = {
       // This carries over any public data from the anonymous session
@@ -686,7 +682,7 @@ async function createNewSession(
     // This carries over any private data from the anonymous session
     let existingPrivateData = {}
     if (args.jwtPayload?.isAnonymous) {
-      const session = await global.sessionConfig.getSession(
+      const session = await (global as any).sessionConfig.getSession(
         args.jwtPayload.handle,
       )
       if (session) {
@@ -694,7 +690,7 @@ async function createNewSession(
           existingPrivateData = JSON.parse(session.privateData)
         }
         // Delete the previous anonymous session
-        await global.sessionConfig.deleteSession(args.jwtPayload.handle)
+        await (global as any).sessionConfig.deleteSession(args.jwtPayload.handle)
       }
     }
 
@@ -705,13 +701,13 @@ async function createNewSession(
 
     const expiresAt = addMinutes(
       new Date(),
-      global.sessionConfig.sessionExpiryMinutes as number,
+      (global as any).sessionConfig.sessionExpiryMinutes as number,
     )
     const handle = generateEssentialSessionHandle()
     const sessionToken = createSessionToken(handle, newPublicData)
     const publicDataToken = createPublicDataToken(newPublicData)
 
-    await global.sessionConfig.createSession({
+    await (global as any).sessionConfig.createSession({
       expiresAt,
       handle,
       userId: newPublicData.userId,
@@ -735,11 +731,11 @@ async function createNewSession(
       antiCSRFToken,
       sessionToken,
     }
-  } else if (global.sessionConfig.method === "advanced") {
+  } else if ((global as any).sessionConfig.method === "advanced") {
     throw new Error("The advanced method is not yet supported")
   } else {
     throw new Error(
-      `Session management method ${global.sessionConfig.method
+      `Session management method ${(global as any).sessionConfig.method
       } is invalid. Supported methods are "essential" and "advanced"`,
     )
   }
@@ -778,12 +774,12 @@ async function refreshSession(
     setPublicDataCookie(req, res, publicDataToken, expiresAt)
     setCSRFCookie(req, res, sessionKernel.antiCSRFToken, expiresAt)
   } else if (
-    global.sessionConfig.method === "essential" &&
+    (global as any).sessionConfig.method === "essential" &&
     "sessionToken" in sessionKernel
   ) {
     const expiresAt = addMinutes(
       new Date(),
-      global.sessionConfig.sessionExpiryMinutes as number,
+      (global as any).sessionConfig.sessionExpiryMinutes as number,
     )
     const publicDataToken = createPublicDataToken(sessionKernel.publicData)
 
@@ -803,23 +799,23 @@ async function refreshSession(
 
     debug("Updating session in db with", { expiresAt })
     if (publicDataChanged) {
-      await global.sessionConfig.updateSession(sessionKernel.handle, {
+      await (global as any).sessionConfig.updateSession(sessionKernel.handle, {
         expiresAt,
         hashedSessionToken: hash256(sessionToken),
         publicData: JSON.stringify(sessionKernel.publicData),
       })
     } else {
-      await global.sessionConfig.updateSession(sessionKernel.handle, {
+      await (global as any).sessionConfig.updateSession(sessionKernel.handle, {
         expiresAt,
       })
     }
-  } else if (global.sessionConfig.method === "advanced") {
+  } else if ((global as any).sessionConfig.method === "advanced") {
     throw new Error("refreshSession() not implemented for advanced method")
   }
 }
 
 export async function getAllSessionHandlesForUser(userId: PublicData["userId"]) {
-  return (await global.sessionConfig.getSessions(userId)).map(
+  return (await (global as any).sessionConfig.getSessions(userId)).map(
     (session: SessionKernel) => session.handle,
   )
 }
@@ -829,7 +825,7 @@ async function syncPubicDataFieldsForUserIfNeeded(
   data: Record<string, unknown>,
 ) {
   const dataToSync: Record<string, unknown> = {}
-    ; global.sessionConfig.publicDataKeysToSyncAcrossSessions?.forEach(
+    ; (global as any).sessionConfig.publicDataKeysToSyncAcrossSessions?.forEach(
       (key: string) => {
         if (data[key]) {
           dataToSync[key] = data[key]
@@ -837,14 +833,14 @@ async function syncPubicDataFieldsForUserIfNeeded(
       },
     )
   if (Object.keys(dataToSync).length) {
-    const sessions = await global.sessionConfig.getSessions(userId)
+    const sessions = await (global as any).sessionConfig.getSessions(userId)
 
     for (const session of sessions) {
       const publicData = JSON.stringify({
         ...(session.publicData ? JSON.parse(session.publicData) : {}),
         ...dataToSync,
       })
-      await global.sessionConfig.updateSession(session.handle, { publicData })
+      await (global as any).sessionConfig.updateSession(session.handle, { publicData })
     }
   }
 }
@@ -858,7 +854,7 @@ async function revokeSession(
   debug("Revoking session", handle)
   if (!anonymous) {
     try {
-      await global.sessionConfig.deleteSession(handle)
+      await (global as any).sessionConfig.deleteSession(handle)
     } catch (error) {
       // Ignore any errors, like if session doesn't exist in DB
     }
@@ -871,14 +867,14 @@ async function revokeSession(
 }
 
 async function revokeAllSessionsForUser(userId: PublicData["userId"]) {
-  let sessionHandles = (await global.sessionConfig.getSessions(userId)).map(
+  let sessionHandles = (await (global as any).sessionConfig.getSessions(userId)).map(
     (session: SessionKernel) => session.handle,
   )
 
   let revoked: string[] = []
   for (const handle of sessionHandles) {
     try {
-      await global.sessionConfig.deleteSession(handle)
+      await (global as any).sessionConfig.deleteSession(handle)
     } catch (error) {
       // Ignore any errors, like if session doesn't exist in DB
     }
@@ -891,7 +887,7 @@ async function getPublicData(sessionKernel: SessionKernel): Promise<PublicData |
   if (sessionKernel.jwtPayload?.publicData) {
     return sessionKernel.jwtPayload?.publicData
   } else {
-    const session = await global.sessionConfig.getSession(
+    const session = await (global as any).sessionConfig.getSession(
       sessionKernel.handle,
     )
     if (!session) {
@@ -906,7 +902,7 @@ async function getPublicData(sessionKernel: SessionKernel): Promise<PublicData |
 }
 
 async function getPrivateData(handle: string): Promise<Record<any, any> | null> {
-  const session = await global.sessionConfig.getSession(handle)
+  const session = await (global as any).sessionConfig.getSession(handle)
   if (session && session.privateData) {
     return JSON.parse(session.privateData) as Record<any, any>
   } else {
@@ -919,7 +915,7 @@ async function setPrivateData(sessionKernel: SessionKernel, data: Record<any, an
   if (existingPrivateData === null) {
     // Anonymous sessions may not exist in the DB yet
     try {
-      await global.sessionConfig.createSession({
+      await (global as any).sessionConfig.createSession({
         handle: sessionKernel.handle,
       })
     } catch (error) { }
@@ -929,7 +925,7 @@ async function setPrivateData(sessionKernel: SessionKernel, data: Record<any, an
     ...existingPrivateData,
     ...data,
   })
-  await global.sessionConfig.updateSession(sessionKernel.handle, {
+  await (global as any).sessionConfig.updateSession(sessionKernel.handle, {
     privateData,
   })
 }
@@ -962,7 +958,7 @@ export async function setPublicDataForUser(userId: PublicData["userId"], data: R
   // Don't allow updating userId
   delete data.userId
 
-  const sessions = await global.sessionConfig.getSessions(userId)
+  const sessions = await (global as any).sessionConfig.getSessions(userId)
   for (const session of sessions) {
     // Merge data
     const publicData = JSON.stringify({
@@ -970,7 +966,7 @@ export async function setPublicDataForUser(userId: PublicData["userId"], data: R
       ...data,
     })
 
-    await global.sessionConfig.updateSession(session.handle, { publicData })
+    await (global as any).sessionConfig.updateSession(session.handle, { publicData })
   }
 }
