+++ packages/blitz/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "@blitzjs/config/tsconfig.library.json",
+  "compilerOptions": {
+    "lib": ["DOM"]
+  },
   "include": ["."],
   "exclude": ["dist", "build", "node_modules"]
 }
